public.json: Add category.featured
Similar to address.preference, but this is set by Azure on categories
shared by all users, while address.preference is set by customers on
their own addresses.  I've tried to reflect that shift by changing the
attribute name from 'preference' to 'featured'.

An example use-case would be to fetch the top N child categories of a
given parent, without having to care about the details behind the
definition of "top".