public.json: Add .active to payment methods
This allows us to return historical payment methods (e.g. so customers
can see which method they used for a past payment), while still
allowing customers to distinguish between methods that can and cannot
be used for new payments.

The permissions around toggling .active will be complicated.  For
example, a customer should be allowed to change creditCard.active from
true to false if they want to disable a particular card.  In that
case, they should also be able to immediately change .active back to
true.  However, they should not be able to re-activate cards that have
expired, or re-activate payroll-deductions if they are no longer an
employee.  The backend may implement whatever restrictions it feels
are appropriate in this respect, and clients should be prepared to
handle assorted 400s when .active-changing requests violate a
backend-imposed restrictoon.