public.json: Add packagedProduct.gtin13 and ?gtin13=... filters
On Thu, Apr 28, 2016 at 02:24:35PM -0700, Tom Peters wrote [1]:
> Rebecca reported that there was talk of UPC/Barcodes being
> searchable.

This makes a lot of sense to me, especially for wholesalers and such
who want to scan a barcode to order more of a product they're running
low on.  The new field and semantics follow [2,3].

Internally, we allow a product to have several barcodes, and I'm fine
supporting all GTIN-13-compatible barcodes for search.  But for
compatibility with schema.org we should just be returning the primary
barcode here.

[1]: https://github.com/azurestandard/website/issues/304
[2]: https://schema.org/Product
[3]: https://schema.org/gtin13