Add propagate flag to PUT /stop/{id}
We currently support a "shift this drop and all later drops on this
trip by $MINUTES (and optionally persist those changes to future trips
and the route)" UI.  You could implement that with PUTs to the stop in
question and all later stops on that trip, but for big trips that
could be a few dozen requests.  And the current backend implementation
preserves inter-stop relationships (drop-trip and drop-route sequence
integers), which we don't want to churn with actions like:

a. Trip 12 has two stops, stop 34 (target noon, sequence 1) and stop 56
   (target 1pm, sequence 2).

2. PUT /stop/34 shifts the stop forward to target 2pm.
   The backend notices that this moves it after stop 56 and swaps the
   sequence numbers (so stop 56, target 1pm, sequence 1 and stop 34,
   target 2pm, sequence 2).

3. PUT /stop/56 shifts the stop forward to target 3pm.
   The backend notices that this moves it after stop 34 and swaps the
   sequence numbers (so stop 34, target 2pm, sequence 1 and stop 56,
   target 3pm, sequence 2).

With this commit, you can:

  PUT /stop/34?propagate=true

and the backend will notice that this does not require resequencing.