public.json: Add GET /orders?inline=... and /order/{id}?inline=...
For folks who don't want to make a separate request to get the
associated payment method.

I considered replacing order.payment with an ID and having inline
support for that too, but that would require splitting the "please
charge this order to ..." modeling from "we charged that order to ..."
modeling.  That's probably a worthwhile change to make, but I'm
punting on it for now.