public.json: Add additional generic payment methods (cash-on-delivery, etc.)
We could have made this change by adding additional types in the
spirit of payrollDeduction, but it seemed less confusing to convert
payrollDeduction to a generic model.  The drawback to this approach is
that we can't document a single enum entry.

Mappings between Azure's legacy payment methods and the new type
slugs:

  Legacy label                   | API slug
  -------------------------------+-----------------
  C.O.D.                         | cash-on-delivery
  Cash Only                      | cash-on-delivery
  Distributor Customer           | net-10
  Hold for Check                 | hold-for-payment
  Money Order or Cashier's Check | net-10
  Net 10 Days                    | net-10
  Net 30 Days                    | net-30
  Pre Pay                        | hold-for-payment

More background for cash-on-delivery and net-N on Wikipedia [1,2].

[1]: https://en.wikipedia.org/wiki/Cash_on_delivery
[2]: https://en.wikipedia.org/wiki/Net_D