public.json: Include breakdown ancestors in packagedProduct.stock
The previous spec wording was unclear, but the current site includes
breakdown ancestors, so we want to mimic that here [1].  You can
almost figure out the ancestor-included values from the direct values
(e.g. if we set 0 stock for the 6 x 14 and 14 oz packaging in the new
doc example), but the API doesn't clarify what can be broken down into
what.

[1]: https://github.com/azurestandard/beehive/issues/1780