public.json: Add 'confirmed' to order.status
I wasn't sure we needed this when I wrote 82da4ce2 (public.json: Add
/orders and /order/{id}, 2014-12-02).  The alternatives are:

* Check to see if the associated stop still exists, because we delete
  stops at confirmation-time [1].
* Check to see if the associated trip is confirmed [2].

But having to check another model to see if your order was accepted is
tiptoeing on the boundary of "this is enough of a per-order property
that it deserves an order status" for me and David McAtee wants it
[3].

See b8a75015 (public.json: Add stop.short-drop and
/stops?short-drop=(true|false), 2016-05-17) for discussion of
"confirmed" vs. "verified".

[1]: $ jq .definitions.stop.description public.json
[2]: $ jq .definitions.trip.properties.confirmed public.json
[3]: https://github.com/azurestandard/website/issues/404#issuecomment-221801391