public.json: Adjust PUT /order/{id} to return a changedOrder model
And add ?save=false support.  This lets folks check the side effects
of an order-save (e.g. maybe they adjusted the shipment from a drop to
parcel-carrier, and want to know if we'll still ship all of their
ordered products via parcel-carrier).

PUTting an order can also change fees and such, but those aren't
listed in .changes because clients can compare the resulting .order
with a previous retrieval.  Comparing order lines with that approach
would require more work up-front to collect the original data, and
more post-PUT work to see if any lines had been removed or changed.

We currently only record line-removals in .changes.  But with the
coming price overhaul, shipping costs are probably going to be
per-line [1] (in which case we'll probably grow a reprice-order-line
orderChange type, or some such).

[1]: On Slack on 2016-03-15:

  13:40 < wking> benjamin: with the planned pricing module, are
    product prices going to be adjusted depending on the target drop
    (e.g. costing more for longer trips), or are we keeping product
    prices fixed and adding shipping fees (like the current per-route
    shipping fee percentage)?
  13:41 < benjamin> It will be by route, rather than drop
  13:42 < wking> but will the calculated fee increase the product
    cost, or be applied as a fee afterward
  13:43 < benjamin> Yes it will increase the product cost. Each
    customer will see unique pricing on the web, based on route and
    many other criteria
  13:43 < wking> ok, so a $10 product on a long route will look like a
    $11 product (or whatever) and not a $10 product with a $1 shipping
    fee
  13:43 < benjamin> yes