public.json: Add updatePackagedProductAudit.size
To allow for non-size distinctions between packaging without having to
split them into separate product families.  The vinegar example is for
CO463, but this change will also be useful for variable-weight
packaging such as MT159 (6 x 6-10 lb turkey).

Using 'size' for this field matches the existing packagedProduct.size
and packagedProductAudit.size.  Now that we're including non-size
information the name isn't as good a match, but it's probably not
worth renaming the existing instances, and it is important to keep the
property names consistent between the related models.