public.json: Split order.fees into its own endpoint
This allows the backend to use more expensive fee-estimation logic
(e.g. hitting parcel-carrier APIs) without bogging down order
manipulations.  We initially expected order-manipulation hitting fee
recalculation to be rare [1], but it turns out to be unacceptably slow
[2], and we don't need the fee information most of the time [3] (or
it's ok if the fee update shows up asynchronously a second or two so
after the change that triggered it).

[1]: https://github.com/azurestandard/website/issues/186#issuecomment-213072160
[2]: https://github.com/azurestandard/website/issues/552#issue-165217215
[3]: https://github.com/azurestandard/website/issues/552#issuecomment-232406614