public.json: JSON fixups for payment-method DELETE
There was already a /payment-method/{id} key for GET.

This method deactivates the method (instead of deleting it), because
we keep old method information around for displaying old invoices.

We will also allow customers to associate inactive (but unexpired)
credit cards with orders to support the "'Save card for future
purchases' false" workflow.