Add packs and unit to packagedProduct definition
On Mon, Aug 29, 2016 at 02:59:11PM -0700, Tom Peters wrote [1]:
> We want to be able to search by `5 lb cheese`.  In order to do this
> we need to add the package size to the search index.
>
> This can be done in Algolia but I believe we set this up via the
> API.

On Mon, Aug 29, 2016 at 03:08:20PM -0700, W. Trevor King wrote [2]:
> In today's meeting David McAtee brought up the "lb" vs. "pound"
> canonicalization issue.  I think we may want to export .packs and
> .unit (with the unabbreviated unit) and add those to the index too
> (although that would mean extending the API spec).

We will set these for packaging where we have a reliable value, and
not set them for packaging that has yet to have it's packs/unit
audited.  If/when we complete our packs/unit audit, we can require
these properties.

This change makes packagedProductAudit largely superfluous. I'd rather
shift packagedProductAudit.audited over as packs-audited or some such
and only return packs-audited to folks with a change-packaged-product
permission.  Then we can drop packagedProductAudit, and roll the
current audit endpoints into the usual GET /packages and a new PATCH
/packaged-product/{id}.  The PATCH could 500 folks trying to change
anything besides unit/packs/name.  But that's a larger change, so
we're punting on it for now.

[1]: https://github.com/azurestandard/beehive/issues/2133#issue-173883186
[2]: https://github.com/azurestandard/beehive/issues/2133#issuecomment-243272972