public.json: Unify storage enums
Having a separate 'storage' enum makes it easy to keep the existing
drop.storage and updateDrop.storage in sync with the new
product.storage (which I've renamed from storage-climate in this
commit to match the drop properties).  Previous drop users should
expect the following value adjustments:

* refregerator (sic) -> chilled
* freezer -> frozen

as well as a new "greenhouse" value.