Registration confirmation also performs login.
This change would allow clients to treat a successful /registration/confirm response as a successful login of the user that was activated. Login would only occur if activation was successful, so it would only work once per token as activation does.

This would allow for a more streamlined registration flow, as the front end will be able to immediately engage the user with links to add a drop, fill out profile information, etc., without them having to sign in first.