public.json: Authenticate successful registrations immediately
When the user provides an email address, check for existing accounts
which use that email address.

* If an inactive account using the email address is found, send an
  inactive-member email pointing the user at customer service.  Return
  a resend token, but do not authenticate the user.

* If an active account using the email address is found, send an
  active-member email with an authenticating token.  Return a resend
  token, but do not authenticate the user until they provide the auth
  token.

* If no account using that email address is found, create a new
  account and immediately authenticate the user.  If they provided an
  email address or SMS number, potentially send a welcome-aboard
  notification.  Any email address will remain unconfirmed until the
  user completes a confirmation cycle.

This allows a malicious user to:

* Mine the API for Azure-user email addresses (by attempting to
  register an email address and checking to see whether they're
  authenticated or not).

* DoS us with lots of registrations (because they don't need a unique,
  functional email address for each successful registration).

But the business decision is that avoiding the email-confirmation
cycle in the new-user case is worth those drawbacks.

Without the confirmation cycle, we drop the registration/confirm
endpoint.  Folks attempting to register over an active account can
present their auth token to the login endpoint.

Also update "one of your associated email addresses" -> "your primary
email address" to match the existing implementation.  We may adjust
that in the future, but we'd handle it in a separate change-set.