Allow sorting drops by distance.
The general `sort` query parameter means new sort fields can be added in
the future.  Since the distance needs to be calculated from a lat + lon
input, a special format is used here.  The pipe is used as the delimiter
so the comma can be preserved for future fields to sort by, so, e.g., we
could do: `?sort=state,-distance(lat:45.45|lon:-121.13)`.

Fixes #179.