person: Add .internal-notes
Our backend has several ways to store notes associated with a given
user:

a. In an append-only array of internal notes with no UI for editing
   existing notes (via the Django comments app).
b. In an internal, editable text field usually used for billing notes.
c. In an editable text field for default shipping instructions, which
   the customer and internal users can see.

This commit exposes (b).  Our immediate use-case is for recording
promotion codes and similar at registration time so we can create
credits for signing bonuses and the like.