Extend address model with personAddress and updatePersonAddress
Adding person to updatePersonAddress solves the undocumented ability to
add ?person= to the query string of an /addresses POST.

With personAddress, a user with permission to view addresses associated
with other people may now use /addresses?filter-person= with multiple
person ids and know the person associated with each address returned.

Adding geo to personAddress, as we need this for the suggested drop
feature. We could potentially move geo into address, as there are
multiple usages of address and geo in tandem. However, that involves
significant changes to other models and is out of scope here.

This also moves towards more consistency with /emails in terms of how we
handle preference: by defaulting the first personAddress preference to
100, and always returning the preference in personAddress.