Update order fees api to include climate info
Despite the fact that dry is a combination of the dry and greenhouse
storage, we've gone with "dry" instead of "dry/greenhouse" for the
shippingClimate.  There was some formal business discussion about
breaking greenhouse out into its own shipping climate [1], but we
didn't record the result of that discussion.  Trevor remembers that
breaking greenhouse out into it's own shipping climate was rejected,
but does not remember whether there was discussion of "dry"
vs. "dry/greenhouse" for the name.

[1]: https://github.com/azurestandard/azure/wiki/Monday-Software-Meeting-20170227#ups-api