public.json: Add stop.drop-shipping-fee
So users can calculate shipping fees will be without changing the
order and hitting /order-fees.  The backend currently hard-codes this
at 8.5% and allows stops to be toggled independently, so I have not
added the new property to updateStop.

The "drop" prefix is from discussion with Casey, where we'd like to
guard against potential future pickup-oriented fee information being
exposted in the API (e.g. if the current
route.backhaul-cost-per-pickup is adjusted to vary by stop instead of
being fixed equal for all stops on a route).