WIP: public.json: Add /person/{id}/prices for requesting per-customer prices
The current retail/wholesale/member categorization is too coarse, and
the business folks want a way to have per-customer deals for customers
(or groups of customers) on packaged products (or groups of packaged
products).  With this commit, we remove the old price-levels and
replace them with a single basePrice for each packaged-product.
Individual customers (or customer service representatives acting on
behalf of customers) can hit the new endpoint to find
packaged-products where they have a custom price that's not the
default.  This commit covers use cases like "all employees get 10% off
everything", "all members of drop 12 get $5 off food->flour
packaged-products", and "customer 34 gets $3 off FL006".

One future use case might be promotional codes, where users
who push a promo code (to an as-yet-unspecified API endpoint) would
pick up the reduced price.

Caching this sort of thing is going to be tricky, because the
parameters that a custom price depends on are unclear.  Based on a dev
meeting with Tom, Jordan, Casey, and David McAtee today, we're
expecting prices returned by this endpoint to generally be valid until
midnight, which we can represent with the usual cache-control HTTP
headers.  It's not clear to me yet if having a custom price returned
to you via the API guarantees you at least that price until the cache
expires.  For example,

1. You hit /person/{id}/prices and are quoted $10 for FL006 (because
   you're a member of drop 12, although the API doesn't tell you
   this).  The default price for FL006 is $12.
2. You leave drop 12.
3. You add an order-line for FL006 to your placed order scheduled to
   be delivered to drop 34.

Can the backend give you the $12 per-unit price on your order-line?
Must it remember that you had a valid-until-midnight $10 quote and
give you that instead?  Will we never adjust custom-price logic on the
backend in ways that might violate a claimed valid-until-midnight
cache statement (this might be tricky if we base prices on the
current, possibly estimated, landed cost)?

The query filtering is by product instead of packaged-product because
users are expected to always want to know the prices (custom or
otherwise) for all packaging associated with a given product.  The
frontend may display this information as a price range (or a
price-per-unit range) even if the customer is not shown a separate
price for each packaging.

FIXME: Will we have any products where we are going to price them by
weight, but the MSRP is fixed (not by weight) or vice versa?  Or will
we always agree with the manufacturer about whether a given
packaged-product should be by weight or not?

FIXME: What values do we want per-unit numbers for?  MSRP-per-unit?
Azure-default-retail-per-unit?  Your-customer-specific-price-per-unit?

FIXME: Would we ever want different wholesale/retail MSRPs?

FIXME: Do we need a way to tag sale items for Algolia filtering?  The
price-level-based on-sale-* tags I'm removing in this commit currently
back https://www.azurestandard.com/shop/tag/on-sale.  What are we
doing without that page now that "what products are on sale?" becomes
a per-customer question?  Do all "deals" count as "sales", or are
"sales" a subset of "deals", or are "sales" and "deals" completely
independent things?