public.json: Add /files and /file/{id}
The way we've been managing product images internally has proven
awkward and buggy.  This commit adds a file API so we can handle
images (and other opaque files) more explicitly and benefit from the
same logging that applies to the rest of our API.

There is no GET support for individual files, because the contents are
available via https://media.azurestandard.com/files/{id}.

There is no PUT support for individual files, because files are
immutable to support more efficient caching.  If you want to replace a
category image, you should:

1. Upload the new image via POST /files, remembering the resulting
   file ID.
2. PUT to /category/{id} with the new file ID in 'image'.
3. DELETE /file/{old-id}.

This commit has no provision for supplying a media type with the
upload, since we don't store file media types on the backend.  The
initial consumer will have to figure out the media type by inspecting
the content.  Currently all of our file consumers use the opaque files
for images, and Thumbor (which the frontend uses for sizing all
images) handles this auto-detection for us.  If we start using File
for non-images, we'll have to add explicit media type handling.

Swagger v2 doesn't have a way to declare non-JSON request bodies [1],
but with OAI v3 (about to be released) we will be able to do that [2].

[1]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#fixed-fields-7
[2]: https://github.com/OAI/OpenAPI-Specification/blob/3.0.0-rc2/versions/3.0.md#requestBodyObject