public.json: Add /packaged-product-images and /packaged-product-image/{id}
Replacing the old packageProduct.images with a more detailed model.
This gives an API for updating packaged-product images via the /files
and /file/{id} endpoints which is similar to the recent category-image
API.

Consumers of packagedProduct models who want image URLs should inline
'image' and construct the URLs from the image names.  For example:

  $ echo '{"images": [{"id": 12, "packaged-product": "FL006", "image": "aa-bb-cc", "order": 3}]}' | jq '[.images[] | "https://media.azurestandard.com/file/" + .image]'
  [
    "https://media.azurestandard.com/files/aa-bb-cc"
  ]