public.json: Add sale endpoints
And grow the sale model so it can stand alone.  This expansion will
allow us to use the public sale API for both the website and the
internal pricing UI.

We may end up making past/future sale information internal; this
commit doesn't have an opinion on what sales are accessible to a given
user.

Including current sales in price objects by default is convenient, but
should probably be spun off into an inline at some point.  However,
tracing inlines from packaged-products through prices to sales is a
bit awkward to talk about, so I've skipped it for now.