public.json: Add limit=none support for /brands
I'm not sure if this is entirely legal Swagger, but it gets the idea
across.

David McAtee wants this because it's more convenient than paging [1].
The brands API is cheap (~340 ms in production from BH-AP-1 with
limit=1 and ~420 ms with limit=250, so in production for from BH-AP-1,
so ~0.32ms per brand) and we only have ~1k brands in production with a
low chance of many more.  So I'm not worried about the load of
unlimited /brands requests.

I want to keep the default limit lower than the cap to make it less
likely that users accidentally submit expensive requests.  If you want
a high-limit or unlimited response, you need to ask for it explicitly.

Potential APIs for explicitly unlimited responses:

a. Pick an absurdly high number (?limit=999999999).
b. Use a sentinal value (?limit=null, ?limit=none, ?limit=unlimited,
   ?limit=-1, ...).

I don't have a personal preference.  (a) would require the least
backend change, but none of those approaches would be particularly
hard to support.  It doesn't seem like much difference from the
frontend side either, since things will likely be breaking all over
the place by the time we attempt to serialize a billion brands in one
response and we'd need to go back to paginating.  David liked
?limit=none [2], so that's what I'm using here.

[1]: https://github.com/azurestandard/beehive/pull/2960#issuecomment-318811267
[2]: https://github.com/azurestandard/beehive/issues/2984#issuecomment-319780972