public.json: Add category.seo-name
The website is currently [1] setting a <title> element [2] based on
hard-coded values [3] like this.  The values landed in the website via
azurestandard/website@a801e764 (Added more titles and descriptions,
2016-08-19) citing:

On Tue, Jul 26, 2016 at 04:15:29PM -0700, Tom Peters wrote [4]:
> ###  5.2  Product Page Meta Title
>
> The product title (brand - product name) is used for this.

and landing titles like:

  Organic Chicken & Animal Feed in Bulk - Organic Farming & Garden
  Supplies

provided by Lee Ladyga [5].  That's a pretty long title; too long for
our sidebar listing sub-categories [6].  But the name in the sidebar
is easier to read if it's not as compact as the short name, which we
want to use for breadcrumbs, URL slugs, and such [7].  So we have
three different size needs:

* A longish form for <title> (seo-name).
* A medium form for the sidebar and other low-context locations
  (name).
* A short form for breadcrumbs, URL slugs, and other high-context
  locations (short-name).

I've gone with 'seo-name' instead of 'long-name' because David expects
the values to be SEO-specific (with too many keywords or odd phrasing
to be worth displaying in other places?):

On Fri, Sep 01, 2017 at 05:28:35PM +0000, David McAtee wrote [8]:
> > if we want a very short name (for the URL and such), a medium name
> > (for the sidebar and such), and a long name (for SEO), then an
> > seo-title (or seo-name?  long-name?) property makes sense.
>
> Yes, I think this is what we need. I like `seo-title` because I
> think we should only use this specifically for that purpose.

I've used 'name' rather than 'title', because this value is covering
the same idea as our existing name and short-name, and I like the
(*-)name pattern.

I've also documented the short-name -> name suggested fallback to keep
up with the new seo-name -> name suggested fallback.

[1]: https://github.com/azurestandard/website/blob/v1.7.8/src/index.html#L8
[2]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/title
[3]: https://github.com/azurestandard/website/blob/v1.7.8/src/app.shop/shop_listing.index.controller.js#L506-L525
[4]: https://github.com/azurestandard/website/issues/654
[5]: https://github.com/azurestandard/website/issues/654#issuecomment-241164075
[6]: https://github.com/azurestandard/website/blob/v1.7.8/src/app.shop/shop_listing.index.html#L84-L87
[7]: https://github.com/azurestandard/website/issues/1738
[8]: https://github.com/azurestandard/api-spec/issues/213#issuecomment-326638890