public.json: Add person.order-place-issue
If a customer has carried a sufficiently large balance for a
sufficiently long time (as determined by the backend), we want to at
least warn users instead of letting them silently checkout (and
continue to grow their balance).  We have a few classes of customers
for this situation:

* Some customers have fairly serious issues, and they should be
  encouraged to contact customer service to resolve the issue.  Until
  then, we will block attempts to place orders.

* Some customers have fairly standard issues, and they should be
  encouraged to settle their balance directly (e.g. via the "Make a
  Payment" button on [1]).  They can also call customer service to
  sort things out that way.  Until their balance is resolved, we will
  block attempts to place orders.

* Some customers have special arrangements with Azure to enable them
  to carry balances while continuing to place new orders.  They should
  be warned about their outstanding balance, but will still be allowed
  to place their order.

The enum slugs all start with "outstanding-balance-" to future-proof
against other cases where a customer should be blocked or warned from
placing all orders.  We don't have those cases now, but using specific
slugs leaves room for them if we get some in the future.

The expected workflow is that when users visit a view that allows them
to place an order, the frontend will pull a fresh person to get the
current value of this property.  If the property is set, the frontend
will display an appropriate warning/error, disable the order-placing
button, etc.  Because customers are likely to resolve these issues out
of band (e.g. via a phone call), frontend warnings should include a
suggestion to refresh the data if the customer considers it stale.

There are also related checks that are per-order (vs. per-customer).
For example. we allow customers to place orders after cutoff (but
before verification) if their drop received short-drop notifications.
Because they are order-specific, they cannot be represented in this
per-person property.  We discussed using PUT /order/{id}?save=false
for this check, returning errors for these outstanding-balance cases.
That would have allowed us to get the currently-implemented per-order
checks at that point as well.  The benefit of the per-customer
property is that if at some point we decide to warn customers earlier
(e.g. at login, or with a site-wide banner), this approach will work
while a per-order approach would require at least an active order.

The new inline tag allows folks to explicitly ask for this property,
because calculating a balance may be more expensive than other person
properties.  The /person entry for inline was missing entirely, and
should have been added for 'email' back in fcf3467 (public.json: Add
GET ...?inline=email to /people and /person/{id}, 2016-08-19, #150).

The current slug pattern is {reason}-{frontend-action-hint}.  I don't
see a need to document that in the API spec itself.  I expect the
frontend will take action for any recognized slug, and show a generic
warning for any unrecognized slug.  Being able to patrially-unpack
unrecognized entries (e.g. if you understood only the reason part)
might allow for more-targeted fallbacks, but I don't expect the
backend to so outpace the frontend to be worried about a generic
warning.  If we did want to support partially-recognized situations,
I'd rather be using a more structured response, like:

  "order-place-issue": {
    "reason": "outstanding-balance",
    "action": "pay"
  }

But again, I think we're fine without that.

[1]: https://www.azurestandard.com/my-account/account-balance